move all inference methods into `trainers` folder
