+++ sbi/inference/trainers/npse/npse.py
