+++ sbi/inference/trainers/snle/snle_a.py
