+++ sbi/inference/trainers/snle/snle_base.py
