+++ sbi/inference/trainers/snpe/__init__.py
