+++ sbi/inference/trainers/snpe/snpe_a.py
