+++ sbi/inference/trainers/snre/bnre.py
