+++ sbi/inference/trainers/snre/snre_a.py
