+++ sbi/inference/trainers/snre/snre_b.py
