+++ sbi/inference/trainers/snre/snre_base.py
