+++ sbi/inference/trainers/snre/snre_c.py
